+++ src/main.rs
@@ -12,7 +12,6 @@ extern crate rustc_span;
 extern crate rustc_stable_hash;
 extern crate stable_mir;
 
-use distributed_verification::kani_path;
 use functions::{clear_rustc_ctx, set_rustc_ctx};
 use rustc_middle::ty::TyCtxt;
 use stable_mir::CrateDef;
@@ -26,50 +25,27 @@ extern crate tracing;
 
 fn main() {
     logger::init();
-    let cli = cli::parse();
-    let kani_path = kani_path();
-    info!(kani_path, ?cli);
-    let mut args = Vec::from(
-        [
-            // the first argument to rustc is unimportant
-            "rustc",
-            "--crate-type=lib",
-            "--cfg=kani",
-            "-Zcrate-attr=feature(register_tool)",
-            "-Zcrate-attr=register_tool(kanitool)",
-            "--sysroot",
-            &kani_path,
-            "-L",
-            &format!("{kani_path}/lib"),
-            "--extern",
-            "kani",
-            "--extern",
-            &format!("noprelude:std={kani_path}/lib/libstd.rlib"),
-            "-Zunstable-options",
-            "-Zalways-encode-mir",
-            "-Zmir-enable-passes=-RemoveStorageMarkers",
-        ]
-        .map(String::from),
-    );
-    args.extend(cli.rustc_args);
+    let run = cli::parse();
 
-    let res = run_with_tcx!(args, |tcx| {
+    let res = run_with_tcx!(run.args, |tcx| {
         use eyre::{Context, Ok};
 
-        let crates = stable_mir::external_crates();
-        dbg!(crates.len(), crates);
-        for krate in stable_mir::find_crates("core") {
-            let fn_defs = krate.fn_defs();
-            dbg!(fn_defs.len());
-            for fn_def in fn_defs {
-                let name = fn_def.name();
-                let attrs = fn_def.all_tool_attrs();
-                // let attrs = fn_def.tool_attrs(&["kanitool".into(), "proof".into()]);
-                if attrs.is_empty() { continue; }
-                let attrs = attrs.iter().map(|attr| attr.as_str()).collect::<Vec<_>>().join(" ");
-                println!("{name}: {attrs:?}");
-            }
-        }
+        // let crates = stable_mir::external_crates();
+        // dbg!(crates.len(), crates);
+        // for krate in stable_mir::find_crates("core") {
+        //     let fn_defs = krate.fn_defs();
+        //     dbg!(fn_defs.len());
+        //     for fn_def in fn_defs {
+        //         let name = fn_def.name();
+        //         let attrs = fn_def.all_tool_attrs();
+        //         // let attrs = fn_def.tool_attrs(&["kanitool".into(), "proof".into()]);
+        //         if attrs.is_empty() {
+        //             continue;
+        //         }
+        //         let attrs = attrs.iter().map(|attr| attr.as_str()).collect::<Vec<_>>().join(" ");
+        //         println!("{name}: {attrs:?}");
+        //     }
+        // }
 
         set_rustc_ctx(tcx);
 
@@ -77,7 +53,7 @@ fn main() {
 
         clear_rustc_ctx();
 
-        let res = || match &cli.json {
+        let res = || match &run.json {
             Some(path) => {
                 if path == "false" {
                     return Ok(());
